requirements.txt
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-PageDown==0.4.0
Flask-SimpleMDE==0.3.0
Flask-Markdown==0.3
gevent==21.1.2
pytest==7.1.2
PyYAML==6.0
SQLAlchemy==1.3.10
WTForms==3.0.1